requirements.txt
Django>=4.2,<7.0
django-ninja-simple-jwt>=0.7.1
django-ninja>=1.5.3
olefile>=0.47
packaging>=23.2
Pillow>=10.0.0
pyparsing>=3.1.0
six>=1.16.0
wheel>=0.41.0
